implementacion del service y controller

+++ src/main/java/com/luchodevs/cashtaxi/service/TrayectoService.java
@@ -2,18 +2,21 @@ package com.luchodevs.cashtaxi.service;
 
 
 import com.luchodevs.cashtaxi.dto.TrayectoDto;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
+
+
 
 public interface TrayectoService {
 
     List<TrayectoDto> listarTrayectos();
 
-    TrayectoDto obtenerPorId(int id);
+    TrayectoDto obtenerPorId(Long id);
 
     TrayectoDto guardarTrayecto(TrayectoDto trayectoDto);
 
-    TrayectoDto actualizarTrayecto(int id, TrayectoDto trayectoDto);
+    TrayectoDto actualizarTrayecto(Long id, TrayectoDto trayectoDto);
 
-    void eliminarTrayecto(int id);
+    void eliminarTrayecto(Long id);
 }
